Move consider-using-namedtuple-or-dataclass to CodeStyle extension
