+++ ChangeLog
@@ -124,9 +124,6 @@ modules are added.
   Closes #4509
   Closes PyCQA/astroid#999
 
-* New checker ``consider-using-namedtuple-or-dataclass``. Emitted when dictionary values
-  can be replaced by namedtuples or dataclass instances.
-
 * New checker ``invalid-all-format``. Emitted when ``__all__`` has an invalid format,
   i.e. isn't a ``tuple`` or ``list``.
