+++ tests/functional/ext/code_style/code_style_consider_using_namedtuple_or_dataclass.rc
@@ -1,0 +1,2 @@
+[MASTER]
+load-plugins=pylint.extensions.code_style
