+++ tests/functional/ext/code_style/code_style_consider_using_namedtuple_or_dataclass.txt
